move promts to file for docker-compose subgen
