+++ generators/docker-compose/index.js
@@ -7,6 +7,7 @@ var generators = require('yeoman-generator'),
     jsyaml = require('js-yaml'),
     pathjs = require('path'),
     util = require('util'),
+    prompts = require('./prompts'),
     scriptBase = require('../generator-base');
 
 var DockerComposeGenerator = generators.Base.extend({});
@@ -81,196 +82,16 @@ module.exports = DockerComposeGenerator.extend({
         }
     },
 
-    _getAppFolders: function (input) {
-        var files = shelljs.ls('-l', this.destinationPath(input));
-        var appsFolders = [];
-
-        files.forEach(function(file) {
-            if(file.isDirectory()) {
-                if( (shelljs.test('-f', input + file.name + '/.yo-rc.json'))
-                    && (shelljs.test('-f', input + file.name + '/src/main/docker/app.yml')) ) {
-                    try {
-                        var fileData = this.fs.readJSON(input + file.name + '/.yo-rc.json');
-                        if(fileData['generator-jhipster'].baseName !== undefined) {
-                            appsFolders.push(file.name.match(/([^\/]*)\/*$/)[1]);
-                        }
-                    } catch(err) {
-                        this.log(chalk.red(file + ': this .yo-rc.json can\'t be read'));
-                    }
-                }
-            }
-        }, this);
-
-        return appsFolders;
-    },
-
     prompting: {
-        askForPath: function() {
-            if (this.regenerate) return;
+        askForPath: prompts.askForPath,
 
-            var done = this.async();
+        askForApps: prompts.askForApps,
 
-            var prompts = [{
-                type: 'input',
-                name: 'directoryPath',
-                message: 'Enter the root directory where your gateway(s) and microservices are located',
-                default: this.directoryPath || '../',
-                validate: function (input) {
-                    var path = this.destinationPath(input);
-                    if(shelljs.test('-d', path)) {
-                        var appsFolders = this._getAppFolders(input);
+        askForClustersMode: prompts.askForClustersMode,
 
-                        if(appsFolders.length === 0) {
-                            return 'No microservice or gateway found in ' + path;
-                        } else {
-                            return true;
-                        }
-                    } else {
-                        return path + ' is not a directory or doesn\'t exist';
-                    }
-                }.bind(this)
-            }];
+        askForElk: prompts.askForElk,
 
-            this.prompt(prompts, function (props) {
-                this.directoryPath = props.directoryPath;
-
-                this.appsFolders = this._getAppFolders(this.directoryPath);
-
-                //Removing registry from appsFolders, using reverse for loop
-                for(var i = this.appsFolders.length - 1; i >= 0; i--) {
-                    if (this.appsFolders[i] === 'jhipster-registry' || this.appsFolders[i] === 'registry') {
-                        this.appsFolders.splice(i,1);
-                    }
-                }
-
-                this.log(chalk.green(this.appsFolders.length + ' applications found at ' + this.destinationPath(this.directoryPath) + '\n'));
-
-                done();
-            }.bind(this));
-        },
-
-        askForApps: function() {
-            if (this.regenerate) return;
-
-            var done = this.async();
-
-            var prompts = [{
-                type: 'checkbox',
-                name: 'chosenApps',
-                message: 'Which applications do you want to include in your Docker Compose configuration?',
-                choices: this.appsFolders,
-                default: this.defaultAppsFolders,
-                validate: function (input) {
-                    if(input.length === 0) {
-                        return 'Please choose at least one application';
-                    } else return true;
-                }
-            }];
-
-            this.prompt(prompts, function (props) {
-                this.appsFolders = props.chosenApps;
-
-                this.appConfigs = [];
-                this.gatewayNb = 0;
-                this.monolithicNb = 0;
-                this.microserviceNb = 0;
-
-                //Loading configs
-                this.appsFolders.forEach(function (appFolder) {
-                    var path = this.destinationPath(this.directoryPath + appFolder +'/.yo-rc.json');
-                    var fileData = this.fs.readJSON(path);
-                    var config = fileData['generator-jhipster'];
-
-                    if(config.applicationType === 'monolith') {
-                        this.monolithicNb++;
-                    } else if(config.applicationType === 'gateway') {
-                        this.gatewayNb++;
-                    } else if(config.applicationType === 'microservice') {
-                        this.microserviceNb++;
-                    }
-
-                    this.portsToBind = this.monolithicNb + this.gatewayNb;
-                    this.appConfigs.push(config);
-                }, this);
-
-                done();
-            }.bind(this));
-        },
-
-        askForClustersMode: function () {
-            if (this.regenerate) return;
-
-            var mongoApps = [];
-            this.appConfigs.forEach(function (appConfig, index) {
-                if(appConfig.prodDatabaseType === 'mongodb') {
-                    mongoApps.push(this.appsFolders[index]);
-                }
-            }, this);
-            if(mongoApps.length===0) return;
-
-            var done = this.async();
-
-            var prompts = [{
-                type: 'checkbox',
-                name: 'clusteredDbApps',
-                message: 'Which applications do you want to use with clustered databases (only available with MongoDB)?',
-                choices: mongoApps,
-                default: this.clusteredDbApps
-            }];
-
-            this.prompt(prompts, function (props) {
-                this.clusteredDbApps = props.clusteredDbApps;
-                for (var i = 0; i < this.appsFolders.length; i++) {
-                    for (var j = 0; j < props.clusteredDbApps.length; j++) {
-                        this.appConfigs[i].clusteredDb = this.appsFolders[i] === props.clusteredDbApps[j];
-                    }
-                }
-
-                done();
-            }.bind(this));
-        },
-
-        askForElk: function() {
-            if (this.regenerate) return;
-
-            var done = this.async();
-
-            var prompts = [{
-                type: 'confirm',
-                name: 'elk',
-                message: 'Do you want the JHipster Console (based on ELK) to monitor your applications?',
-                default: this.useElk && true
-            }];
-
-            this.prompt(prompts, function(props) {
-                this.useElk = props.elk;
-                done();
-            }.bind(this));
-        },
-
-        askForAdminPassword: function() {
-            if (this.regenerate) return;
-
-            var done = this.async();
-
-            var prompts = [{
-                type: 'input',
-                name: 'adminPassword',
-                message: 'Enter the admin password used to secure the JHipster Registry',
-                default: 'admin',
-                validate: function (input) {
-                    if(input.length < 5) {
-                        return 'The password must have at least 5 characters';
-                    } else return true;
-                }
-            }];
-
-            this.prompt(prompts, function(props) {
-                this.adminPassword = props.adminPassword;
-                done();
-            }.bind(this));
-        }
-    },
+        askForAdminPassword: prompts.askForAdminPassword,
 
     configuring: {
         insight: function () {
